WA-89 Moved crypto from core to login for modlist
