+++ src/core/login-core.js
@@ -2,7 +2,6 @@ import _ from 'lodash';
 import {throwStatus} from '../util/express';
 
 import jwt from 'jwt-simple';
-import crypto from 'crypto';
 const {knex} = require('../util/database').connect();
 require('../init-env-variables');
 
@@ -137,12 +136,6 @@ export function modList() {
     if (_.isEmpty(rows)) {
       return throwStatus(404, 'No moderator list');
     } else {
-      for (var i = 0; i < rows.length; i++) {
-        const decipher = crypto.createDecipher('aes192', process.env.CRYPTO_PASSWORD)
-        var email = decipher.update(rows[i].email, 'hex', 'utf8');
-        email += decipher.final('utf8')
-        rows[i].email = email
-      }
       return rows;
     }
   });
